requirements.txt
pip~=22.0.4
setuptools~=59.1.1
psycopg2==2.9.2
python-dotenv==0.19.2
Django==3.2.9
djangorestframework==3.12.4
celery==5.2.3
redis==4.2.0
eventlet==0.33.0
django-celery-beat==2.2.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0